2025/12/04/guardrails-agentic-devops/python/requirements.txt
requests>=2.31.0
pyyaml>=6.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
